requirements.txt
alembic==1.12.1
anyio==3.6.2
APScheduler==3.6.3
autopep8==2.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
flake8==5.0.4
greenlet==3.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==4.2.0
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.5
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
zipp==3.15.0
